Added paragraph_parser
Paragaph_parser.py is a the code to implement parsing the article by paragraph, combine paragraphs to then summarize and lastly compare that summary to trigger, or not to trigger, the claim activation that is hoped to be implemented in sprint 3 / later on. Early worries of inconsistent and improper summarization and leverage cosine similarity (sentence comparison) to gain good results. Much more testing needed.