Sentence comparison beginning with sentence_weighing.py
This file will stem and remove stop words from each sentence OR paragraph and then compare each sentence/paragraph to the title/summary and do that for a variety of algorithms. the values from each algorithm will be averaged out to compute a score per entry and will be evaluated at the end for k_para_display and claim_qery branches.